+++ src/reactor_common.h
@@ -11,7 +11,7 @@
 typedef struct io_state_t
 {
   struct iovec vec[2];
-  size_t vec_sz;
+  int32_t vec_sz;
   command_t cmd;
 } io_state_t;
 
@@ -21,7 +21,7 @@ typedef struct write_state_t
   struct iovec vec_extra[3];
   command_t cmd_extra;
   int32_t length;
-  size_t vec_extra_sz;
+  int32_t vec_extra_sz;
 } write_state_t;
 
 typedef struct reactor_context_t
@@ -50,7 +50,7 @@ void *handle_io (void *arg);
 void *dispatch_event_loop (void *arg);
 
 status_t write_state_write_wrapper (int socket_fd, struct iovec *vec,
-                                    size_t *vec_sz);
+                                    int *vec_sz);
 status_t write_state_write (int socket_fd, write_state_t *write_state);
 
 status_t push_job (reactor_context_t *rctr_ctx, void *arg,
